+++ src/main/java/task_use_cases/TaskDeletionUseCase.java
@@ -1,4 +1,4 @@
-package use_cases;
+package task_use_cases;
 
 import entities.StudentUser;
 import entities.Course;
